fix: rename file to kebab case
